Fix linting errors, add consumer_test.go
